feature/FZ9BF-27-PR-DietList refactoring...
