+++ src/components/Content/CentralCarousel/Carousel.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import Carousel from 'nuka-carousel';
 
-import slide_1 from '../../assets/carousel_slide_1.jpg';
-import slide_2 from '../../assets/carousel_slide_2.jpg';
-import slide_3 from '../../assets/carousel_slide_3.jpg';
-import slide_4 from '../../assets/carousel_slide_4.jpg';
+import slide_1 from '../../../assets/carousel_slide_1.jpg';
+import slide_2 from '../../../assets/carousel_slide_2.jpg';
+import slide_3 from '../../../assets/carousel_slide_3.jpg';
+import slide_4 from '../../../assets/carousel_slide_4.jpg';
 
 
 const SectionCarousel = () => {
