+++ src/components/Content/CentralCarousel/index.js
