+++ src/components/Favourites/Favourites.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import Sidebar from '../Sidebar';
+import Sidebar from '../Sidebar/index';
 import {withStyles} from '@material-ui/core/styles';
 
 const styles = theme => ({
